+++ src/main/java/pl/edu/pwr/pwrinspace/poliwrocket/Controller/Unused/QDPullController.java
@@ -1,19 +1,19 @@
-package pl.edu.pwr.pwrinspace.poliwrocket.Controller;
+package pl.edu.pwr.pwrinspace.poliwrocket.Controller.Unused;
 
 import com.jfoenix.controls.JFXButton;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
-import javafx.scene.control.Label;
 import javafx.scene.layout.AnchorPane;
+import pl.edu.pwr.pwrinspace.poliwrocket.Controller.BaseCommandsController;
 import pl.edu.pwr.pwrinspace.poliwrocket.Model.Command.ICommand;
 import pl.edu.pwr.pwrinspace.poliwrocket.Model.SerialPort.SerialPortManager;
 
 import java.util.Comparator;
 import java.util.stream.Collectors;
 
-public class TanwaCloseController extends BaseCommandsController {
+public class QDPullController extends BaseCommandsController {
     @FXML
     private AnchorPane mainPanel;
 
@@ -26,12 +26,12 @@ public class TanwaCloseController extends BaseCommandsController {
         tileHashMap.clear();
         indicatorHashMap.clear();
 
-        int initY = 35;
+        int initY = 10;
         int offsetY = 51;
 
         for (ICommand command : this.commands.stream().sorted(Comparator.comparing(ICommand::getCommandDescription)).collect(Collectors.toList())) {
 
-            JFXButton button = new JFXButton("CLOSE");
+            JFXButton button = new JFXButton("PULL");
 
             button.setId(command.getCommandTriggerKey());
             button.setLayoutX(5);
