+++ src/main/java/pl/edu/pwr/pwrinspace/poliwrocket/Controller/Unused/TanwaCloseController.java
@@ -1,62 +1,48 @@
-package pl.edu.pwr.pwrinspace.poliwrocket.Controller;
+package pl.edu.pwr.pwrinspace.poliwrocket.Controller.Unused;
 
 import com.jfoenix.controls.JFXButton;
-import com.jfoenix.controls.JFXTextField;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
-import javafx.scene.control.Label;
 import javafx.scene.layout.AnchorPane;
+import pl.edu.pwr.pwrinspace.poliwrocket.Controller.BaseCommandsController;
 import pl.edu.pwr.pwrinspace.poliwrocket.Model.Command.ICommand;
 import pl.edu.pwr.pwrinspace.poliwrocket.Model.SerialPort.SerialPortManager;
 
 import java.util.Comparator;
 import java.util.stream.Collectors;
 
-public class TanwaOpenController extends BaseCommandsController {
-
+public class TanwaCloseController extends BaseCommandsController {
     @FXML
     private AnchorPane mainPanel;
 
     @Override
     protected void buildVisualizationMap() {
         mainPanel.getChildren().removeIf(node ->
-                labelHashMap.containsValue(node)
-                        || buttonHashMap.containsValue(node)
+                buttonHashMap.containsValue(node)
         );
 
         tileHashMap.clear();
         indicatorHashMap.clear();
-        labelHashMap.clear();
 
-        int initYLabel = 45;
         int initY = 35;
         int offsetY = 51;
 
         for (ICommand command : this.commands.stream().sorted(Comparator.comparing(ICommand::getCommandDescription)).collect(Collectors.toList())) {
-            Label label = new Label(command.getCommandDescription());
 
-            JFXButton button = new JFXButton("OPEN");
-
-            label.setLayoutX(14);
-            label.setLayoutY(initYLabel);
-            label.setPrefHeight(18);
-            label.setPrefWidth(180);
+            JFXButton button = new JFXButton("CLOSE");
 
             button.setId(command.getCommandTriggerKey());
-            button.setLayoutX(180);
+            button.setLayoutX(5);
             button.setLayoutY(initY);
             button.setPrefHeight(34);
             button.setPrefWidth(72);
 
-            mainPanel.getChildren().add(label);
             mainPanel.getChildren().add(button);
 
-            labelHashMap.put(command.getCommandTriggerKey(), label);
             buttonHashMap.put(command.getCommandTriggerKey(), button);
 
-            initYLabel += offsetY;
             initY += offsetY;
 
         }
