+++ src/main/java/pl/edu/pwr/pwrinspace/poliwrocket/Controller/Unused/TanwaOpenController.java
@@ -1,4 +1,4 @@
-package pl.edu.pwr.pwrinspace.poliwrocket.Controller;
+package pl.edu.pwr.pwrinspace.poliwrocket.Controller.Unused;
 
 import com.jfoenix.controls.JFXButton;
 import javafx.event.ActionEvent;
@@ -7,13 +7,15 @@ import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.layout.AnchorPane;
+import pl.edu.pwr.pwrinspace.poliwrocket.Controller.BaseCommandsController;
 import pl.edu.pwr.pwrinspace.poliwrocket.Model.Command.ICommand;
 import pl.edu.pwr.pwrinspace.poliwrocket.Model.SerialPort.SerialPortManager;
 
 import java.util.Comparator;
 import java.util.stream.Collectors;
 
-public class QDPushController extends BaseCommandsController {
+public class TanwaOpenController extends BaseCommandsController {
+
     @FXML
     private AnchorPane mainPanel;
 
@@ -28,14 +30,14 @@ public class QDPushController extends BaseCommandsController {
         indicatorHashMap.clear();
         labelHashMap.clear();
 
-        int initYLabel = 20;
-        int initY = 10;
+        int initYLabel = 45;
+        int initY = 35;
         int offsetY = 51;
 
         for (ICommand command : this.commands.stream().sorted(Comparator.comparing(ICommand::getCommandDescription)).collect(Collectors.toList())) {
             Label label = new Label(command.getCommandDescription());
 
-            JFXButton button = new JFXButton("PUSH");
+            JFXButton button = new JFXButton("OPEN");
 
             label.setLayoutX(14);
             label.setLayoutY(initYLabel);
